requirements.txt
python-decouple==3.3
ipdb==0.13.2
ipython==7.13.0